Separated drawers from the canvas
