+++ src/canvas.h
@@ -92,23 +92,24 @@ struct IsometricCanvas {
 
   // Empty section with only beams
   void renderBeamSection(const int64_t, const int64_t, const uint8_t);
-
-  // This obscure typedef allows to create a member function pointer array
-  // (ouch) to render different block types without a switch case
-  typedef void (IsometricCanvas::*drawer)(const uint32_t, const uint32_t,
-                                          const NBT &, const Colors::Block *);
-
-  // The default block type, hardcoded
-  void drawFull(const uint32_t, const uint32_t, const NBT &,
-                const Colors::Block *);
-
-  // The other block types are loaded at compile-time from the `blocktypes.def`
-  // file, with some macro manipulation
-#define DEFINETYPE(STRING, CALLBACK)                                           \
-  void CALLBACK(const uint32_t, const uint32_t, const NBT &,                   \
-                const Colors::Block *);
-#include "./blocktypes.def"
-#undef DEFINETYPE
 };
+
+inline void blend(uint8_t *const destination, const uint8_t *const source) {
+  if (!source[PALPHA])
+    return;
+
+  if (destination[PALPHA] == 0 || source[PALPHA] == 255) {
+    memcpy(destination, source, BYTESPERPIXEL);
+    return;
+  }
+#define BLEND(ca, aa, cb)                                                      \
+  uint8_t(((size_t(ca) * size_t(aa)) + (size_t(255 - aa) * size_t(cb))) / 255)
+  destination[0] = BLEND(source[0], source[PALPHA], destination[0]);
+  destination[1] = BLEND(source[1], source[PALPHA], destination[1]);
+  destination[2] = BLEND(source[2], source[PALPHA], destination[2]);
+  destination[PALPHA] +=
+      (size_t(source[PALPHA]) * size_t(255 - destination[PALPHA])) / 255;
+#undef BLEND
+}
 
 #endif
